Move selection related utilities to utils/selection.js.
