+++ packages/ckeditor5-table/src/commands/selectcolumncommand.js
@@ -11,7 +11,7 @@ import Command from '@ckeditor/ckeditor5-core/src/command';
 
 import TableWalker from '../tablewalker';
 import { findAncestor } from './utils';
-import { getSelectionAffectedTableCells } from '../utils/common';
+import { getSelectionAffectedTableCells } from '../utils/selection';
 
 /**
  * The select column command.
