+++ packages/ckeditor5-table/src/tablenavigation.js
@@ -10,12 +10,12 @@
 import TableSelection from './tableselection';
 import TableWalker from './tablewalker';
 import { findAncestor } from './commands/utils';
-import { getSelectedTableCells, getTableCellsContainingSelection } from './utils/common';
 
 import Plugin from '@ckeditor/ckeditor5-core/src/plugin';
 import Rect from '@ckeditor/ckeditor5-utils/src/dom/rect';
 import priorities from '@ckeditor/ckeditor5-utils/src/priorities';
 import { keyCodes } from '@ckeditor/ckeditor5-utils/src/keyboard';
+import { getSelectedTableCells, getTableCellsContainingSelection } from './utils/selection';
 
 /**
  * This plugin enables keyboard navigation for tables.
