+++ packages/ckeditor5-table/tests/manual/tablemocking.js
@@ -14,8 +14,8 @@ import { diffString } from 'json-diff';
 import { debounce } from 'lodash-es';
 import ArticlePluginSet from '@ckeditor/ckeditor5-core/tests/_utils/articlepluginset';
 import TableWalker from '../../src/tablewalker';
-import { getSelectionAffectedTableCells } from '../../src/utils/common';
 import { findAncestor } from '../../src/commands/utils';
+import { getSelectionAffectedTableCells } from '../../src/utils/selection';
 
 ClassicEditor
 	.create( document.querySelector( '#editor' ), {
